フルスクリーンモードでメインウィンドウが閉じられても吉里吉里が終了しない不具合を修正 (DirectXのバグ？メッセージキューが空のときにPostQuitMessageを呼んでもWM_QUITが飛んでこない。他のメッセージと一緒にPostするとOK)
git-svn-id: svn://localhost/kirikiri2/trunk@785 131d44d2-bf85-024a-9602-49a0ea93a7a9

+++ kirikiri2/src/core/base/win32/SysInitImpl.cpp
@@ -1211,10 +1211,21 @@ void TVPAfterSystemUninit()
 
 
 //---------------------------------------------------------------------------
+bool TVPTerminated = false;
+//---------------------------------------------------------------------------
 void TVPTerminateAsync()
 {
 	// do "A"synchronous temination of application
+	TVPTerminated = true;
+
+	// posting dummy message will prevent "missing WM_QUIT bug" in DirectDraw framework.
+	if(TVPMainForm)
+		::PostMessage(TVPMainForm->Handle, WM_USER+0x31/*dummy msg*/, 0, 0);
+
 	Application->Terminate();
+
+	if(TVPMainForm)
+		::PostMessage(TVPMainForm->Handle, WM_USER+0x31/*dummy msg*/, 0, 0);
 }
 //---------------------------------------------------------------------------
 void TVPTerminateSync()
